requirements.txt
# Core dependencies
flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# RAG and LLM
langchain==0.1.0
langchain-community==0.0.10
openai==1.6.1

# Vector DB and Embeddings
chromadb==0.4.22
sentence-transformers==2.2.2

# Document processing
pypdf==3.17.4
markdown==3.5.1
beautifulsoup4==4.12.2

# Utilities
numpy==1.24.3
requests==2.31.0
